refactor(backend): deepen skill package composition
